[INTERNAL] refactor filesystem related code
